pull save message into its own component
